+++ src/utils/resources.ts
@@ -1,33 +1,7 @@
-import { ThemeClass, ThemeName } from "../components/interfaces";
-
-const autoTheme = "calcite-theme-auto";
-const darkTheme = "calcite-theme-dark";
-const lightTheme = "calcite-theme-light";
-
-interface Theme {
-  name: ThemeName;
-  className: ThemeClass;
-}
-
-export const THEMES: Theme[] = [
-  {
-    name: "light",
-    className: lightTheme
-  },
-  {
-    name: "dark",
-    className: darkTheme
-  },
-  {
-    name: "auto",
-    className: autoTheme
-  }
-];
-
 export const CSS_UTILITY = {
-  autoTheme,
-  darkTheme,
-  lightTheme,
+  autoTheme: "calcite-theme-auto",
+  darkTheme: "calcite-theme-dark",
+  lightTheme: "calcite-theme-light",
   rtl: "calcite--rtl"
 };
